Moved many cdf files to lib
