+++ packages/cdf/index.js
@@ -1,8 +1,8 @@
-const { Cdf } = require('./cdf');
-const { Pdf } = require('./pdf');
-const { Samples } = require('./samples');
-const { ContinuousDistributionCombination } = require('./continuousDistributionCombination');
-const { ContinuousDistribution } = require('./continuousDistribution');
+const { Cdf } = require('./lib/cdf');
+const { Pdf } = require('./lib/pdf');
+const { Samples } = require('./lib/samples');
+const { ContinuousDistributionCombination } = require('./lib/continuousDistributionCombination');
+const { ContinuousDistribution } = require('./lib/continuousDistribution');
 
 module.exports = {
   Pdf,
