+++ packages/cdf/lib/continuousDistribution.js
