+++ packages/cdf/lib/continuousDistribution.spec.js
