+++ packages/cdf/lib/continuousDistributionCombination.js
